backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic[email]
supabase
python-jose[cryptography]
passlib[bcrypt]
python-multipart
